\docType{package}
\name{rk.subset-package}
\alias{rk.subset}
\alias{rk.subset-package}
\title{The rk.subset Package}
\description{
  RKWard GUI to define subsets of data objects.
}
\details{
  \tabular{ll}{ Package: \tab rk.subset\cr Type: \tab
  Package\cr Version: \tab 0.01-2\cr Date: \tab
  2013-02-24\cr Depends: \tab rkward (>= 0.5.6)\cr
  Enhances: \tab rkward\cr License: \tab GPL (>= 3)\cr
  LazyLoad: \tab yes\cr URL: \tab http://rkward.sf.net\cr }

  RKWard GUI to define subsets of data objects
}
\author{
  Meik Michalke \email{meik.michalke@hhu.de}
}
\keyword{package}

